Data.Default.Class

Documentation

This module defines a class for types with a default value. It also defines Default instances for the types Int, Int8, Int16, Int32, Int64, Word, Word8, Word16, Word32, Word64, Integer, Float, Double, Ratio, Complex, CShort, CUShort, CInt, CUInt, CLong, CULong, CLLong, CULLong, CPtrdiff, CSize, CSigAtomic, CIntPtr, CUIntPtr, CIntMax, CUIntMax, CClock, CTime, CUSeconds, CSUSeconds, CFloat, CDouble, '(->)', IO, Maybe, '()', '[]', Ordering, Any, All, Last, First, Sum, Product, Endo, Dual, and tuples.

class Default a where

A class for types with a default value.

Methods

def :: a

The default value for this type.

def :: (Generic a, GDefault (Rep a)) => a

The default value for this type.